Started validating maps, opening fd

+++ ft_mlx_pixel_put.c
@@ -6,7 +6,7 @@
 /*   By: emende <emende@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/26 21:01:30 by emende            #+#    #+#             */
-/*   Updated: 2022/03/28 14:58:29 by emende           ###   ########.fr       */
+/*   Updated: 2022/04/01 19:52:53 by emende           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,5 +17,5 @@ void	ft_mlx_pixel_put(t_data *data, int x, int y, int color)
 	char	*dst;
 
 	dst = data->addr + (y * data->line_len + x * (data->bpp / 8));
-	*(unsigned int *)dst = color;
+	*(unsigned int *)dst = (unsigned int) color;
 }
